chore: MotionMeasurement 관련 파일 분리 #6
